+++ tests/unit/StoreGatewayTest.php
@@ -1,22 +1,28 @@
 <?php
 
+use Carbon\Carbon;
+use Codeception\Test\Unit;
 use Faker\Factory;
 use Faker\Generator;
 use Foodsharing\Modules\Core\DBConstants\Store\CooperationStatus;
 use Foodsharing\Modules\Core\DBConstants\StoreTeam\MembershipStatus;
+use Foodsharing\Modules\Store\DTO\CreateStoreData;
 use Foodsharing\Modules\Store\StoreGateway;
 use Foodsharing\Modules\Store\TeamStatus;
 
-class StoreGatewayTest extends \Codeception\Test\Unit
+class StoreGatewayTest extends Unit
 {
 	protected UnitTester $tester;
 	private Generator $faker;
 	private StoreGateway $gateway;
 
-	private $store;
-	private $foodsaver;
-	private $region;
+	private array $store;
+	private array $foodsaver;
+	private array $region;
 
+	/**
+	 * @throws Exception
+	 */
 	private function storeData($status = 'none'): array
 	{
 		return [
@@ -35,13 +41,13 @@ class StoreGatewayTest extends \Codeception\Test\Unit
 			'str' => $this->store['str'],
 			'plz' => $this->store['plz'],
 			'stadt' => $this->store['stadt'],
-			'added' => (new \DateTime($this->store['added']))->format('Y-m-d'),
+			'added' => (new DateTime($this->store['added']))->format('Y-m-d'),
 			'verantwortlich' => ($status === 'team') ? 0 : null,
 			'active' => ($status === 'team') ? 1 : null,
 		];
 	}
 
-	protected function _before()
+	protected function _before(): void
 	{
 		$this->gateway = $this->tester->get(StoreGateway::class);
 		$this->region = $this->tester->createRegion();
@@ -50,31 +56,58 @@ class StoreGatewayTest extends \Codeception\Test\Unit
 		$this->faker = Factory::create('de_DE');
 	}
 
-	public function testIsInTeam()
+	public function testAddNewStore(): void
 	{
-		$this->assertEquals(TeamStatus::NoMember,
+		$storeDTO = new CreateStoreData();
+		$storeDTO->name = 'StoreGatewayTestbetrieb';
+		$storeDTO->regionId = 1567;
+		$storeDTO->lat = 51.5367827;
+		$storeDTO->lon = 9.9258967;
+		$storeDTO->str = 'Bahnhofsplatz 1';
+		$storeDTO->zip = '37073';
+		$storeDTO->city = 'Göttingen';
+		$storeDTO->publicInfo = 'Testeintrag im Feld öffentliche Information';
+		$storeDTO->createdAt = Carbon::now();
+		$storeDTO->updatedAt = Carbon::now();
+
+		$storeId = $this->gateway->addStore($storeDTO);
+
+		$this->assertIsInt($storeId);
+		$this->assertTrue($storeId !== 0);
+	}
+
+	public function testIsInTeam(): void
+	{
+		$this->assertEquals(
+			TeamStatus::NoMember,
 			$this->gateway->getUserTeamStatus($this->foodsaver['id'], $this->store['id'])
 		);
 
 		$this->tester->addStoreTeam($this->store['id'], $this->foodsaver['id']);
-		$this->assertEquals(TeamStatus::Member,
+		$this->assertEquals(
+			TeamStatus::Member,
 			$this->gateway->getUserTeamStatus($this->foodsaver['id'], $this->store['id'])
 		);
 
 		$coordinator = $this->tester->createStoreCoordinator();
 		$this->tester->addStoreTeam($this->store['id'], $coordinator['id'], true);
-		$this->assertEquals(TeamStatus::Coordinator,
+		$this->assertEquals(
+			TeamStatus::Coordinator,
 			$this->gateway->getUserTeamStatus($coordinator['id'], $this->store['id'])
 		);
 
 		$waiter = $this->tester->createFoodsaver();
 		$this->tester->addStoreTeam($this->store['id'], $waiter['id'], false, true);
-		$this->assertEquals(TeamStatus::WaitingList,
+		$this->assertEquals(
+			TeamStatus::WaitingList,
 			$this->gateway->getUserTeamStatus($waiter['id'], $this->store['id'])
 		);
 	}
 
-	public function testListStoresForFoodsaver()
+	/**
+	 * @throws Exception
+	 */
+	public function testListStoresForFoodsaver(): void
 	{
 		$this->assertEquals(
 			[
@@ -101,7 +134,7 @@ class StoreGatewayTest extends \Codeception\Test\Unit
 		);
 	}
 
-	public function testUpdateStoreRegion()
+	public function testUpdateStoreRegion(): void
 	{
 		$newRegion = $this->tester->createRegion();
 
@@ -110,39 +143,72 @@ class StoreGatewayTest extends \Codeception\Test\Unit
 		$this->tester->seeInDatabase('fs_betrieb', ['bezirk_id' => $newRegion['id'], 'id' => $this->store['id']]);
 	}
 
-	public function testGetNoTeamConversation()
+	public function testGetNoTeamConversation(): void
 	{
 		$conversationId = $this->gateway->getBetriebConversation($this->store['id']);
 
 		$this->tester->assertEquals(0, $conversationId);
 	}
 
-	public function testGetNoSpringerConversation()
+	public function testGetNoSpringerConversation(): void
 	{
 		$conversationId = $this->gateway->getBetriebConversation($this->store['id'], true);
 
 		$this->tester->assertEquals(0, $conversationId);
 	}
 
-	public function testFoodsaverRelatedStoreMembershipStatus()
+	public function testFoodsaverRelatedStoreMembershipStatus(): void
 	{
-		$store1 = $this->tester->createStore($this->region['id'], null, null, ['betrieb_status_id' => CooperationStatus::COOPERATION_ESTABLISHED]);
-		$store2 = $this->tester->createStore($this->region['id'], null, null, ['betrieb_status_id' => CooperationStatus::COOPERATION_ESTABLISHED]);
-		$store3 = $this->tester->createStore($this->region['id'], null, null, ['betrieb_status_id' => CooperationStatus::COOPERATION_ESTABLISHED]);
-		$store4 = $this->tester->createStore($this->region['id'], null, null, ['betrieb_status_id' => CooperationStatus::COOPERATION_ESTABLISHED]);
-		$store5 = $this->tester->createStore($this->region['id'], null, null, ['betrieb_status_id' => CooperationStatus::DOES_NOT_WANT_TO_WORK_WITH_US]);
+		$store1 = $this->tester->createStore(
+			$this->region['id'], null, null, ['betrieb_status_id' => CooperationStatus::COOPERATION_ESTABLISHED]
+		);
+		$store2 = $this->tester->createStore(
+			$this->region['id'], null, null, ['betrieb_status_id' => CooperationStatus::COOPERATION_ESTABLISHED]
+		);
+		$store3 = $this->tester->createStore(
+			$this->region['id'], null, null, ['betrieb_status_id' => CooperationStatus::COOPERATION_ESTABLISHED]
+		);
+		$store4 = $this->tester->createStore(
+			$this->region['id'], null, null, ['betrieb_status_id' => CooperationStatus::COOPERATION_ESTABLISHED]
+		);
+		$store5 = $this->tester->createStore(
+			$this->region['id'], null, null, ['betrieb_status_id' => CooperationStatus::DOES_NOT_WANT_TO_WORK_WITH_US]
+		);
 
 		$this->tester->addStoreTeam($store1['id'], $this->foodsaver['id'], true, false, true); // Test coordinator
-		$this->tester->addStoreTeam($store2['id'], $this->foodsaver['id'], false, true, true); // Test waiting for membership (JUMPER) - Not shown
-		$this->tester->addStoreTeam($store3['id'], $this->foodsaver['id'], false, false, true); // Test membership (MEMBER)
-		$this->tester->addStoreTeam($store4['id'], $this->foodsaver['id'], false, false, false); // Test open request confirmed (Pending request) - Not shown
-		$this->tester->addStoreTeam($store5['id'], $this->foodsaver['id'], false, false, false); // Test open request confirmed (Pending request) - Not shown
-		$this->tester->addStoreTeam($store2['id'], $this->foodsaver['id'] + 1, false, false, true); // Test open request confirmed (Pending request)
+		$this->tester->addStoreTeam(
+			$store2['id'], $this->foodsaver['id'], false, true, true
+		); // Test waiting for membership (JUMPER) - Not shown
+		$this->tester->addStoreTeam(
+			$store3['id'], $this->foodsaver['id'], false, false, true
+		); // Test membership (MEMBER)
+		$this->tester->addStoreTeam(
+			$store4['id'], $this->foodsaver['id'], false, false, false
+		); // Test open request confirmed (Pending request) - Not shown
+		$this->tester->addStoreTeam(
+			$store5['id'], $this->foodsaver['id'], false, false, false
+		); // Test open request confirmed (Pending request) - Not shown
+		$this->tester->addStoreTeam(
+			$store2['id'], $this->foodsaver['id'] + 1, false, false, true
+		); // Test open request confirmed (Pending request)
 
-		$expectation = [['store_id' => $store1['id'], 'store_name' => $store1['name'], 'managing' => 1, 'membership_status' => MembershipStatus::MEMBER],
-						['store_id' => $store3['id'], 'store_name' => $store3['name'], 'managing' => 0, 'membership_status' => MembershipStatus::MEMBER],
-						['store_id' => $store2['id'], 'store_name' => $store2['name'], 'managing' => 0, 'membership_status' => MembershipStatus::JUMPER],
-						['store_id' => $store4['id'], 'store_name' => $store4['name'], 'managing' => 0, 'membership_status' => MembershipStatus::APPLIED_FOR_TEAM]
+		$expectation = [
+			[
+				'store_id' => $store1['id'], 'store_name' => $store1['name'], 'managing' => 1,
+				'membership_status' => MembershipStatus::MEMBER,
+			],
+			[
+				'store_id' => $store3['id'], 'store_name' => $store3['name'], 'managing' => 0,
+				'membership_status' => MembershipStatus::MEMBER,
+			],
+			[
+				'store_id' => $store2['id'], 'store_name' => $store2['name'], 'managing' => 0,
+				'membership_status' => MembershipStatus::JUMPER,
+			],
+			[
+				'store_id' => $store4['id'], 'store_name' => $store4['name'], 'managing' => 0,
+				'membership_status' => MembershipStatus::APPLIED_FOR_TEAM,
+			],
 		];
 		usort($expectation, function ($a, $b) {
 			if ($a['managing'] == $b['managing']) {
@@ -168,7 +234,13 @@ class StoreGatewayTest extends \Codeception\Test\Unit
 				return 1;
 			}
 		});
-		$result = $this->gateway->listAllStoreTeamMembershipsForFoodsaver($this->foodsaver['id'], [CooperationStatus::UNCLEAR, CooperationStatus::NO_CONTACT, CooperationStatus::IN_NEGOTIATION, CooperationStatus::COOPERATION_STARTING, CooperationStatus::COOPERATION_ESTABLISHED, CooperationStatus::PERMANENTLY_CLOSED]);
+		$result = $this->gateway->listAllStoreTeamMembershipsForFoodsaver(
+			$this->foodsaver['id'], [
+				CooperationStatus::UNCLEAR, CooperationStatus::NO_CONTACT, CooperationStatus::IN_NEGOTIATION,
+				CooperationStatus::COOPERATION_STARTING, CooperationStatus::COOPERATION_ESTABLISHED,
+				CooperationStatus::PERMANENTLY_CLOSED,
+			]
+		);
 		$this->assertEquals(4, count($result));
 		$this->assertEquals($expectation[0]['store_id'], $result[0]->store->id);
 		$this->assertEquals($expectation[0]['store_name'], $result[0]->store->name);
@@ -191,16 +263,28 @@ class StoreGatewayTest extends \Codeception\Test\Unit
 		$this->assertEquals($expectation[3]['membership_status'], $result[3]->membershipStatus);
 	}
 
-	public function testStoreCooperationFilterForFoodsaverRelatedStoreMembershipsByStatusNotWantToWorkWithUs()
+	public function testStoreCooperationFilterForFoodsaverRelatedStoreMembershipsByStatusNotWantToWorkWithUs(): void
 	{
-		$store2 = $this->tester->createStore($this->region['id'], null, null, ['betrieb_status_id' => CooperationStatus::COOPERATION_ESTABLISHED]);
-		$store5 = $this->tester->createStore($this->region['id'], null, null, ['betrieb_status_id' => CooperationStatus::DOES_NOT_WANT_TO_WORK_WITH_US]);
+		$store2 = $this->tester->createStore(
+			$this->region['id'], null, null, ['betrieb_status_id' => CooperationStatus::COOPERATION_ESTABLISHED]
+		);
+		$store5 = $this->tester->createStore(
+			$this->region['id'], null, null, ['betrieb_status_id' => CooperationStatus::DOES_NOT_WANT_TO_WORK_WITH_US]
+		);
 
-		$this->tester->addStoreTeam($store5['id'], $this->foodsaver['id'], false, false, false); // Test open request confirmed (Pending request) - Not shown
-		$this->tester->addStoreTeam($store2['id'], $this->foodsaver['id'] + 1, false, false, true); // Test open request confirmed (Pending request)
-		$this->tester->addStoreTeam($store5['id'], $this->foodsaver['id'] + 1, false, false, true); // Test open request confirmed (Pending request)
+		$this->tester->addStoreTeam(
+			$store5['id'], $this->foodsaver['id'], false, false, false
+		); // Test open request confirmed (Pending request) - Not shown
+		$this->tester->addStoreTeam(
+			$store2['id'], $this->foodsaver['id'] + 1, false, false, true
+		); // Test open request confirmed (Pending request)
+		$this->tester->addStoreTeam(
+			$store5['id'], $this->foodsaver['id'] + 1, false, false, true
+		); // Test open request confirmed (Pending request)
 
-		$result = $this->gateway->listAllStoreTeamMembershipsForFoodsaver($this->foodsaver['id'], [CooperationStatus::DOES_NOT_WANT_TO_WORK_WITH_US]);
+		$result = $this->gateway->listAllStoreTeamMembershipsForFoodsaver(
+			$this->foodsaver['id'], [CooperationStatus::DOES_NOT_WANT_TO_WORK_WITH_US]
+		);
 		$this->assertEquals(1, count($result));
 		$this->assertEquals($store5['id'], $result[0]->store->id);
 		$this->assertEquals($store5['name'], $result[0]->store->name);